Don't overwrite an existing description file.

+++ AFQ/api/group.py
@@ -208,8 +208,9 @@ class GroupAFQ(object):
 
         pl_desc_file = op.join(self.afq_path, 'dataset_description.json')
 
-        with open(pl_desc_file, 'w') as outfile:
-            json.dump(pipeline_description, outfile)
+        if not op.exists(pl_desc_file):
+            with open(pl_desc_file, 'w') as outfile:
+                json.dump(pipeline_description, outfile)
 
         self.subjects = bids_layout.get(return_type='id', target='subject')
         if not len(self.subjects):
@@ -670,7 +671,7 @@ class GroupAFQ(object):
         view : str
             Which view to display. Can be one of sagittal, coronal, or axial.
         direc : str
-            Which direction to views. Can be one of left, right, top, bottom, 
+            Which direction to views. Can be one of left, right, top, bottom,
             front, back
         slice_pos : float, or None
             If float, indicates the fractional position along the
